+++ sdk/src/rhesis/sdk/models/__init__.py
@@ -1,5 +1,5 @@
 from rhesis.sdk.models.providers.gemini import GeminiLLM
-from rhesis.sdk.models.providers.openai_provider import OpenAILLM
+from rhesis.sdk.models.providers.openai import OpenAILLM
 from rhesis.sdk.models.providers.rhesis_provider import RhesisLLM
 
 __all__ = ["RhesisLLM", "GeminiLLM", "OpenAILLM"]
